Mise a jour pour l'intégration de l'identification par mot de passe et email , modification des entity pour permettre la mise en place des dates des updates et create automatiquement, suppression de la table typeUser qui n'est plus utiles aux vu des rôles dans la table fos-user
