+++ app/config/security.yml
@@ -11,7 +11,7 @@ security:
     # http://symfony.com/doc/current/book/security.html#where-do-users-come-from-user-providers
     providers:
         fos_userbundle:
-            id: fos_user.user_provider.username
+            id: fos_user.user_provider.username_email
     firewalls:
         # disables authentication for assets and the profiler, adapt it according to your needs
         dev:
